Выносит в отдельный компонент promo-film
