+++ public/index.html
@@ -7,6 +7,7 @@
   <meta name="robots" content="noindex, nofollow">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="stylesheet" href="css/main.min.css">
+  <link rel="stylesheet" href="css/dev-styles.css">
 </head>
 
 <body>
@@ -53,7 +54,7 @@
     </div>
 
   <div id="root"></div>
-  
+
   <script src="./bundle.js"></script>
 </body>
 </html>
