+++ src/components/loading-screen/loading-screen.jsx
@@ -3,10 +3,8 @@ import React from 'react';
 const LoadingScreen = () => {
   return (
 
-    <div className="user-page">
-      <div className="sign-in user-page__content">
-        <h1>Loading...</h1>
-      </div>
+    <div className="loading-screen">
+      <h1>Loading...</h1>
     </div>
   );
 };
