+++ src/components/main-page/main-page-view.jsx
@@ -2,65 +2,15 @@ import React from 'react';
 import CardsList from '../cards-list/cards-list';
 import mainPageViewProps from '../main-page/main-page-view.prop';
 import GenresList from '../genres-list/genres-list';
-import User from '../user/user';
+import PromoFilm from '../promo-film/promo-film';
 
 
 const MainPageView = (props) => {
-  const {promoFilm, filmGenre, initialCount} = props;
-  const {name, genre, posterImage, backgroundImage, released} = promoFilm;
+  const {filmGenre, initialCount} = props;
 
   return <>
 
-    <section className="movie-card">
-      <div className="movie-card__bg">
-        <img src={backgroundImage} alt={name} />
-      </div>
-
-      <h1 className="visually-hidden">WTW</h1>
-
-      <header className="page-header movie-card__head">
-        <div className="logo">
-          <a className="logo__link">
-            <span className="logo__letter logo__letter--1">W</span>
-            <span className="logo__letter logo__letter--2">T</span>
-            <span className="logo__letter logo__letter--3">W</span>
-          </a>
-        </div>
-
-        <User />
-      </header>
-
-      <div className="movie-card__wrap">
-        <div className="movie-card__info">
-          <div className="movie-card__poster">
-            <img src={posterImage} alt={name} width="218" height="327" />
-          </div>
-
-          <div className="movie-card__desc">
-            <h2 className="movie-card__title">{name}</h2>
-            <p className="movie-card__meta">
-              <span className="movie-card__genre">{genre}</span>
-              <span className="movie-card__year">{released}</span>
-            </p>
-
-            <div className="movie-card__buttons">
-              <button className="btn btn--play movie-card__button" type="button">
-                <svg viewBox="0 0 19 19" width="19" height="19">
-                  <use xlinkHref="#play-s"></use>
-                </svg>
-                <span>Play</span>
-              </button>
-              <button className="btn btn--list movie-card__button" type="button">
-                <svg viewBox="0 0 19 20" width="19" height="20">
-                  <use xlinkHref="#add"></use>
-                </svg>
-                <span>My list</span>
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </section>
+    <PromoFilm />
 
     <div className="page-content">
       <section className="catalog">
