+++ src/components/private-route/private-route.jsx
@@ -1,19 +1,17 @@
 import React from 'react';
-import {useSelector} from 'react-redux';
 import {Redirect, Route} from 'react-router';
+import useAuthtorization from '../../api/use-authtorization';
 import privateRouteProps from './private-route.prop';
 
 const PrivateRoute = (props) => {
   const {render, path, exact} = props;
-
-  const authorizationStatus = useSelector((state) => state.authorizationStatus.id);
 
   return (
     <Route
       path={path}
       exact={exact}
       render={(routeProps) => {
-        return (authorizationStatus? render(routeProps) : <Redirect to={`/login`} />);
+        return (useAuthtorization() ? render(routeProps) : <Redirect to={`/login`} />);
       }}
     />
   );
