+++ src/components/promo-film/promo-film-view.prop.js
@@ -1,0 +1,13 @@
+import PropTypes from 'prop-types';
+
+const promoFIlmViewProps = {
+  promoFilm: PropTypes.shape({
+    name: PropTypes.string.isRequired,
+    genre: PropTypes.string.isRequired,
+    released: PropTypes.number.isRequired,
+    posterImage: PropTypes.string.isRequired,
+    backgroundImage: PropTypes.string.isRequired,
+  }).isRequired,
+};
+
+export default promoFIlmViewProps;
